Generate a bunch of varians & types for item_component and move them to a new file.
